sql/catalog/catalogkv: introduce package for writing to catalog kvs
Release note: None
