+++ pkg/sql/grant_revoke.go
@@ -13,6 +13,7 @@ package sql
 import (
 	"context"
 
+	"github.com/cockroachdb/cockroach/pkg/sql/catalog/catalogkv"
 	"github.com/cockroachdb/cockroach/pkg/sql/privilege"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/sqlbase"
@@ -145,7 +146,7 @@ func (n *changePrivilegesNode) startExec(params runParams) error {
 			if err := d.Validate(); err != nil {
 				return err
 			}
-			if err := WriteDescToBatch(
+			if err := catalogkv.WriteDescToBatch(
 				ctx,
 				p.extendedEvalCtx.Tracing.KVTracingEnabled(),
 				p.ExecCfg().Settings,
