+++ pkg/sql/temporary_schema.go
@@ -25,6 +25,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/server/telemetry"
 	"github.com/cockroachdb/cockroach/pkg/settings"
 	"github.com/cockroachdb/cockroach/pkg/settings/cluster"
+	"github.com/cockroachdb/cockroach/pkg/sql/catalog/catalogkv"
 	"github.com/cockroachdb/cockroach/pkg/sql/schema"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/sessiondata"
@@ -83,7 +84,7 @@ var (
 )
 
 func createTempSchema(params runParams, sKey sqlbase.DescriptorKey) (sqlbase.ID, error) {
-	id, err := GenerateUniqueDescID(params.ctx, params.p.ExecCfg().DB, params.p.ExecCfg().Codec)
+	id, err := catalogkv.GenerateUniqueDescID(params.ctx, params.p.ExecCfg().DB, params.p.ExecCfg().Codec)
 	if err != nil {
 		return sqlbase.InvalidID, err
 	}
@@ -140,7 +141,7 @@ func temporarySchemaSessionID(scName string) (bool, ClusterWideID, error) {
 func getTemporaryObjectNames(
 	ctx context.Context, txn *kv.Txn, codec keys.SQLCodec, dbID sqlbase.ID, tempSchemaName string,
 ) (TableNames, error) {
-	dbDesc, err := MustGetDatabaseDescByID(ctx, txn, codec, dbID)
+	dbDesc, err := catalogkv.MustGetDatabaseDescByID(ctx, txn, codec, dbID)
 	if err != nil {
 		return nil, err
 	}
@@ -169,7 +170,7 @@ func cleanupSessionTempObjects(
 	return db.Txn(ctx, func(ctx context.Context, txn *kv.Txn) error {
 		// We are going to read all database descriptor IDs, then for each database
 		// we will drop all the objects under the temporary schema.
-		dbIDs, err := GetAllDatabaseDescriptorIDs(ctx, txn, codec)
+		dbIDs, err := catalogkv.GetAllDatabaseDescriptorIDs(ctx, txn, codec)
 		if err != nil {
 			return err
 		}
@@ -476,7 +477,7 @@ func (c *TemporaryObjectCleaner) doTemporaryObjectCleanup(
 	var dbIDs []sqlbase.ID
 	if err := retryFunc(ctx, func() error {
 		var err error
-		dbIDs, err = GetAllDatabaseDescriptorIDs(ctx, txn, c.codec)
+		dbIDs, err = catalogkv.GetAllDatabaseDescriptorIDs(ctx, txn, c.codec)
 		return err
 	}); err != nil {
 		return err
